==> rmds/02-data-cleaning.Rmd <==
---
title: "more data cleaning"
author: "Lucky Mehra"
date: "9/17/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load-packages}
if(!require(pacman)){
  install.packages("pacman")
}

pacman::p_load(tidyverse,
               qdapRegex)
```

```{r more-cleaning}
sweetpotato <- read_csv(here::here('data/clean-data', 'sweetpotato.csv')) %>% 
  mutate(yield_wf_num_1 = if_else(!is.na(yield_wf_petite_1),
                                 yield_wf_num_1 + yield_wf_petite_1,
                                 yield_wf_num_1),
         
         yield_weedy_num_1 = if_else(!is.na(yield_weedy_petite_1),
                                     yield_weedy_num_1 + yield_weedy_petite_1,
                                     yield_weedy_num_1)) %>% 
  dplyr::select(-yield_wf_petite_1, -yield_weedy_petite_1) %>% 
  rowwise() %>% 
  mutate(yield_wf_marketable = sum(c(yield_wf_num_1, yield_wf_jumbo), na.rm = TRUE),
         yield_wf_total = sum(c(yield_wf_marketable, yield_wf_canner), na.rm = TRUE),
         yield_weedy_marketable = sum(c(yield_weedy_num_1, yield_weedy_jumbo), na.rm = TRUE),
         yield_weedy_total = sum(c(yield_weedy_marketable, yield_weedy_canner), na.rm = TRUE))

# calculate marketable and total yield loss
yieldloss <- sweetpotato %>% 
  mutate(marketable_loss = ((yield_wf_marketable - yield_weedy_marketable)/yield_wf_marketable)*100,
         
         total_loss = ((yield_wf_total - yield_weedy_total)/yield_wf_total)*100)

# clean some variables for consistency
yieldloss <- yieldloss %>% 
  mutate(location_information = if_else(location_information == "Horticultural Crops Research Station",
                                        "Horticultural Crops Research Station, Clinton, NC",
                                        location_information)) %>% 
  mutate(location_information = str_replace_all(location_information, c("-" = ",", "Georgetown" = "Georgetown,")),
         location = ex_city_state(location_information),
         location = str_replace(location, ", ", "_"),
         soil_type = str_replace(soil_type, "loamy sand", "loamy_sand"),
         variety = str_to_lower(variety)) %>% 
  relocate(year, location, variety, primary_weed, soil_type,
           irrigation_type, irrigation_type_present, pi, marketable_loss,
           total_loss) 

yieldloss %>% 
  write_csv(here::here('data/clean-data', 'yieldloss.csv'))
```
